Consolidated PingsPage and PingsLoading page to remidy the issue causing FAQ to load twice. Turns out, doing the fragment transaction actually caused all the fragments in the viewpager to reload!
