--- app/src/main/res/layout/fragment_pings_loading.xml
@@ -1,32 +1,0 @@
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:orientation="vertical"
-    tools:context=".PingsLoadingFragment"
-    android:padding="20dp">
-
-    <TextView
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="@string/loadingPings"
-        android:id="@+id/signInDialogText"
-        android:layout_gravity="center_horizontal"
-        android:layout_margin="10dp" />
-
-    <ProgressBar
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:id="@+id/progressBar"
-        android:layout_gravity="center_horizontal"
-        android:visibility="visible" />
-
-    <Button
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="@string/retryConnection"
-        android:id="@+id/btnRetry"
-        android:layout_gravity="center_horizontal"
-        android:visibility="invisible" />
-
-</LinearLayout>
